requirements.txt
easyocr~=1.7.2
pillow~=11.1.0
pyperclip~=1.9.0
keyboard~=0.13.5
numpy~=2.2.2